requirements.txt
streamlit>=1.33
Pillow>=9
